[WIP; major refactoring]
