+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [ 
     "compute", 
-    "ctensor",
+    "ctensor", "rvn",
     "rvnllm",
     "utils/gguf_loader",
 ]
@@ -18,13 +18,18 @@ repository = "https://github.com/rvnllm/rvnllm"
 
 [workspace.dependencies]
 clap = { version = "4.5.37", features = ["derive"] }
-anyhow = "1.0.98"
-log = "0.4.27"
-env_logger = "0.11.8"
+anyhow = "1.0.98":
 tokio = { version = "1.44.2", features = ["full"] }
 tracing-subscriber = "0.3.19"
 memmap2 = "0.9.5"
 byteorder = "1.5"
+once_cell = "1.21.3"
+rayon = "1.10.0"
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+log = "0.4.27"
+env_logger = "0.11.8"
+bytemuck = "1.23"
 
 [profile.mindebug-dev]
 inherits = "dev"
