+++ rvn/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "rvn"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
