+++ rvnllm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rvn"
+name = "rvnllm"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -7,17 +7,20 @@ homepage.workspace = true
 license.workspace = true
 repository.workspace = true
 
+[lib]
+path = "src/lib.rs"
+
 [dependencies]
 anyhow = { workspace = true }
 memmap2 = { workspace = true }
 byteorder = { workspace = true }
+once_cell = { workspace = true }
+clap = { worksace = true }
+log = { workspace = true }
+env_logger = { workspace = } 
+bytemuck = { } 
+serde = { }
+serde_json = { }
+rayon = { }
 ctensor = { path = "../ctensor" }
 compute = { path = "../compute" }
-once_cell = "1.21.3"
-clap = { workspace = true }
-log = "0.4.27"
-env_logger = "0.11.8"
-bytemuck = "1.23"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-rayon = "1.10.0"
