+++ rvnllm/src/commands.rs
@@ -1,0 +1,12 @@
+// the core dispatcher
+pub mod dispatch;
+// the various supporting commands
+pub mod diff;
+pub mod dump;
+pub mod forward;
+pub mod profile;
+pub mod info;
+pub mod list;
+pub mod watch;
+pub mod validate;
+pub mod analyze;
