+++ rvnllm/src/commands/info.rs
@@ -1,0 +1,4 @@
+pub fn run_info() {
+    println!("run_info")
+}
+
