+++ rvnllm/src/lib.rs
@@ -1,1 +1,5 @@
 pub mod parser;
+pub mod commands;
+
+pub use parser::load_model;
+
