+++ rvnllm/src/parser.rs
@@ -1,0 +1,3 @@
+pub fn load_model() {
+    println!("loading the model");
+}
